Committing the production code for providing liquidity
This commit adds the production code which continuously records all data
from an exchange (in Record_Data/). This data can be accumulated and then
used to train machine learning trading strategies. One of these strategies,
based on the mid-spread, is included in the TraderBot/ directory. This bot
will use this strategy to provide liquidity to a market and make a small
profit from each trade on average.

Co-authored-by: Ciaran Hughes <iamciaran2@gmail.com>
Co-authored-by: Rian Hughes   <rian.hughes@phys.ox.ac.uk>

Why:
To upload our production code to a version control for future use, future
development of features, and perseverance.